Adjust main and add the pipeline in a folder Change the main so its consistent with different zenfiles
